working module/router combo, needs refactoring
